requirements/requirements.txt
Django==3.1.12
django-environ==0.4.5
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
pytest-pythonpath==0.7.3
